requirements.txt
streamlittensorflow-cpunumpyPillow